Move profile to presentation
