--- app/src/main/java/com/example/traders/ErrorEvent.kt
@@ -1,3 +1,0 @@
-package com.example.traders
-
-data class ErrorEvent(val e: Exception)
