+++ app/src/main/java/com/example/traders/presentation/ErrorEvent.kt
@@ -1,0 +1,3 @@
+package com.example.traders.presentation
+
+data class ErrorEvent(val e: Exception)
