+++ app/src/main/java/com/example/traders/presentation/MainActivity.kt
@@ -1,19 +1,17 @@
-package com.example.traders
+package com.example.traders.presentation
 
 import android.animation.Animator
 import android.animation.AnimatorListenerAdapter
 import android.os.Bundle
-import android.util.Log
 import android.view.View
 import androidx.appcompat.app.AppCompatActivity
 import androidx.core.view.forEach
 import androidx.navigation.fragment.NavHostFragment
 import androidx.navigation.ui.setupWithNavController
-import com.example.traders.utils.roundFormatBigDecimal
+import com.example.traders.R
 import com.example.traders.network.webSocket.BinanceWSClient
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import dagger.hilt.android.AndroidEntryPoint
-import java.math.BigDecimal
 import javax.inject.Inject
 
 @AndroidEntryPoint
