+++ app/src/main/java/com/example/traders/presentation/cryptoDetailsScreen/descriptionTab/CryptoDescriptionViewModel.kt
@@ -1,7 +1,7 @@
 package com.example.traders.presentation.cryptoDetailsScreen.descriptionTab
 
 import androidx.lifecycle.viewModelScope
-import com.example.traders.BaseViewModel
+import com.example.traders.presentation.BaseViewModel
 import com.example.traders.network.repository.CryptoRepository
 import com.example.traders.database.FixedCryptoList
 import dagger.hilt.android.lifecycle.HiltViewModel
