+++ app/src/main/java/com/example/traders/presentation/cryptoDetailsScreen/priceStatisticsTab/CryptoPriceStatisticsViewModel.kt
@@ -1,7 +1,7 @@
 package com.example.traders.presentation.cryptoDetailsScreen.priceStatisticsTab
 
 import androidx.lifecycle.*
-import com.example.traders.BaseViewModel
+import com.example.traders.presentation.BaseViewModel
 import com.example.traders.network.repository.CryptoRepository
 import com.example.traders.database.FixedCryptoList
 import com.example.traders.network.models.cryptoStatsData.CryptoStatistics
