+++ app/src/main/java/com/example/traders/presentation/dialogs/buyDialog/BuyDialogEvent.kt
@@ -1,6 +1,6 @@
 package com.example.traders.presentation.dialogs.buyDialog
 
-import com.example.traders.profile.portfolio.TransactionInfo
+import com.example.traders.presentation.profile.portfolio.TransactionInfo
 
 sealed class BuyDialogEvent {
     data class Dismiss(val transactionInfo: TransactionInfo): BuyDialogEvent()
