+++ app/src/main/java/com/example/traders/presentation/dialogs/buyDialog/BuyDialogViewModel.kt
@@ -4,12 +4,12 @@ import android.os.Build
 import androidx.annotation.RequiresApi
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.ViewModelProvider
-import com.example.traders.BaseViewModel
+import com.example.traders.presentation.BaseViewModel
 import com.example.traders.database.Crypto
 import com.example.traders.database.TransactionType
 import com.example.traders.presentation.dialogs.DialogValidation
 import com.example.traders.presentation.dialogs.DialogValidationMessage
-import com.example.traders.profile.portfolio.TransactionInfo
+import com.example.traders.presentation.profile.portfolio.TransactionInfo
 import com.example.traders.network.repository.CryptoRepository
 import com.example.traders.utils.roundNum
 import com.example.traders.database.FixedCryptoList
