+++ app/src/main/java/com/example/traders/presentation/dialogs/sellDialog/SellDialogEvent.kt
@@ -1,6 +1,6 @@
 package com.example.traders.presentation.dialogs.sellDialog
 
-import com.example.traders.profile.portfolio.TransactionInfo
+import com.example.traders.presentation.profile.portfolio.TransactionInfo
 
 sealed class SellDialogEvent {
     data class Dismiss(val transactionInfo: TransactionInfo): SellDialogEvent()
