+++ app/src/main/java/com/example/traders/presentation/profile/ProfileFragment.kt
@@ -1,14 +1,14 @@
-package com.example.traders.profile
+package com.example.traders.presentation.profile
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.viewpager2.widget.ViewPager2
-import com.example.traders.BaseFragment
+import com.example.traders.presentation.BaseFragment
 import com.example.traders.R
 import com.example.traders.databinding.FragmentUserProfileBinding
-import com.example.traders.profile.adapters.UserViewPagerAdapter
+import com.example.traders.presentation.profile.adapters.UserViewPagerAdapter
 import com.google.android.material.tabs.TabLayout
 import com.google.android.material.tabs.TabLayoutMediator
 import dagger.hilt.android.AndroidEntryPoint
