+++ app/src/main/java/com/example/traders/presentation/profile/history/HistoryState.kt
@@ -1,0 +1,5 @@
+package com.example.traders.presentation.profile.history
+
+data class HistoryState(
+    val isEmpty: Boolean
+)
