+++ app/src/main/java/com/example/traders/presentation/profile/portfolio/PortfolioState.kt
@@ -1,4 +1,4 @@
-package com.example.traders.profile.portfolio
+package com.example.traders.presentation.profile.portfolio
 
 import com.example.traders.database.Crypto
 import com.example.traders.network.models.CryptoTicker
