--- app/src/main/java/com/example/traders/profile/history/HistoryState.kt
@@ -1,5 +1,0 @@
-package com.example.traders.profile.history
-
-data class HistoryState(
-    val isEmpty: Boolean
-)
