+++ app/src/main/res/layout/fragment_portfolio.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".profile.ProfileFragment"
+    tools:context=".presentation.profile.ProfileFragment"
     android:background="@color/very_light_gray">
 
     <androidx.core.widget.NestedScrollView
