+++ app/src/main/res/navigation/profile.xml
@@ -8,7 +8,7 @@
 
     <fragment
         android:id="@+id/userProfileFragment"
-        android:name="com.example.traders.profile.ProfileFragment"
+        android:name="com.example.traders.presentation.profile.ProfileFragment"
         android:label="fragment_user_profile"
         tools:layout="@layout/fragment_portfolio" >
         <action
