+++ app/src/test/java/com/example/traders/profile/portfolio/PortfolioViewModelTest.kt
@@ -8,6 +8,8 @@ import com.example.traders.TestDispatcherRule
 import com.example.traders.database.Crypto
 import com.example.traders.network.models.CryptoTicker
 import com.example.traders.network.repository.CryptoRepository
+import com.example.traders.presentation.profile.portfolio.CryptoInUsd
+import com.example.traders.presentation.profile.portfolio.PortfolioViewModel
 import com.example.traders.utils.roundNum
 import com.github.mikephil.charting.data.PieData
 import com.github.mikephil.charting.data.PieEntry
